refactor: setAxiosWithConfig
